Added many single-allocation test files (mostly script generated - may be worth checking upper and lower ends).
Added partial fix to seg faulting problem by not checking for a buddy when in the freeing for loop if the current
order is the highest (where no buddy is possible).

Current iteration still has seg faults in some cases - in particular one of the test cases provided. All single
allocations work as expected.